Allow custom organization name for config profiles
Allow us to set the organization referenced in various configuration
profile via a command-line flag or environment variable.

This commit also removes the ability to set the SCEP Subject via an
argument in service.NewService. The only caller of this function never
set it to anything other than empty string, causing the default to be
used, so it was redundant.